Port cropping tool to new architecture
The cropping tool, like the other tools, is now a seperate class.
There are no changes in functionality, but a small check is added
to determine if the stack is part of a TrackEM2 project. Currently
the backend only allows such projects to be cropped. Therefore, if
the stack is no TrackEM2 stack, the user is displayed a message
that states cropping is not yet possible with that configuration.
This should change in the future to allow cropping for non-TrackEM2
projects, too.