Add input boxes for x, y, width and height coordinates of cropping box
The input boxes allow for direct manipulation of the cropping box in terms
of size and position. This commit changes the coordinates of the cropping
box in a way that now left is always on the left, top always on top and so
on. This makes coordinate handling easier and more robust.