Add custom manager.py command to Django application to add messages to DB
With the help of this "addmessage" command one is able to add a message for
a particular user to the data base through Django. This can be done with
the help of manage.py:

python manage.py addmessage <user_id> <title> <text> <action>

If one is not in Django's virtualenv (e.g. when using a cron job) you can
call it like this:

cd /path/to/django/project && /path/to/django/env/bin/python manage.py addmessage ...

This only works when the database credentials have been set up in settings.py.