Cropping tool: do the actual cropping asynchronously
As cropping requests could take some time due to various reasons (e.g. size
of sub-stack, server load) they should be processed in the background. This
is implemented using the Python module "multiprocessing" which unfortunately
requires access to std.in and std.out -- a behavior that doesn't fit very
well with the use of WSGI. Like mentioned in the INSTALL document a small
configuration change of WSGI prior to 3.0 is needed for this to work.