Add migration that sets up PostGIS representation of skeleton edges
This migration expects the "postgis" extension to be enabled in each
CATMAID database. To do this, please connect as a superuser to your
database and set it up. For instance, if you are on Ubuntu, you should
be able to do the following:

  sudo su postgres -c psql
  \c catmaid
  CREATE EXTENSION postgis;
  \q

Replace "catmaid" in the second line with whatever your actual CATMAID
database name is.

With this done, the added migration sets up a PostGIS edge table that
keeps all edges between treenodes as well as the ID of the child node of
each edge. For root nodes a zero-length edge is stored, to simplify
spatial queries on the front end.

This migration also adds triggers for INSERT, UPDATE and DELETE commands
on the treenode table.