SVG overlay: pin active node before update, if virtual
Virtual nodes don't survive a node update, but will be re-created.
Non-virtual active nodes are fetched from the database, no matter if
they are in the field of view or not. If the active node is virtual,
it is now pinned in the front-end so that it survives an update. This is
required for functionality like interpolation, that requires the active
node to be present in the node dictionary.