Add a preliminary version of abutting connector nodes
A new option ins the tracing section of the settings widget allows users
to create abutting connectors instead of synaptic ones. If this option
is used, nodes will be linked with an "abutting" relation to a
connector. The back-end query for the tracing data now also returns a
list of all nodes that are linked to a connector with another relation
than "presynaptic_to" and "postsynaptic_to". Currently, all nodes that
only have such "other" links are treated as abutting. All other ones are
synaptic connectors. This is likely to change in the future to be more
explicit about other relation types.

Abutting connectors links are displayed with a green color.

This adds a new required relation 'abutting' to the tracing back-end. So
after updating an instance with this commit, a user with admin
permissions should either open all tracing projects in the front-end to
add this relation or use the management command
"catmaid_setup_tracing_for_project"

See catmaid/CATMAID#896