Upgrade Django from v1.6.10 to v1.7.7
Django 1.8 was released beginning of April 2015. This caused Django 1.6
to become unsupported. To be able to receive further security (and
general) updated, this upgrade is performed.

Django 1.7 comes with many changes and some of them require action on
our side. A detailed view on those can be found here:

  https://docs.djangoproject.com/en/1.7/releases/1.7/

The biggest change is that Django now comes with its own migration
system. Therefore, South is removed and a new migration history is
started. All existing custom SQL (like table inheritance, triggers,
indices, etc.) is now done with the new migration system.

This commit fixes all the required things to be able to update in the
first place. For CATMAID, this is in particular:

- Change post_syncdb signal to post_migrate
- Upgrade django-taggit from v0.11.2 to v0.14.0 for Django 1.7 support
- Prevent automatic auto-discover for admin app. This is required for
  django-adminplus to work with Django 1.7.
- Upgrade jsonfield from v1.0.0 to v1.0.3 to include a south migration
  fix.
- Make sure TEMPLATE_DIRS is a tuple.
- Update use of create_superuser management command.