Skeleton projection layer: support display of multiple skeletons
A skeleton projection layer can now subscribe to an arbitrary source to
get an updated list of skeletons. For every skeleton in this list a
projection is displayed. The default source is the active skeleton,
which is effectively what the previous implementation did. However, now
other sources can be selected in the settings widget. So for instance a
Selection Table can provide a list of Skeletons. Through its own
subscriptions, the Selection Table could additionally provide a list
which dynamically includes the active node.

Since there is only one active node, the reference node for additional
skeletons has to be guessed. This is done by finding the first node
closest to the current section.

Closes catmaid/CATMAID#1258