History tables: add history synchronization
With the help of a new PL/pgSQL function, a live table and a history
table can now be synchronized. This is useful, for instance, when
history tables have been disabled and should then become enabled again.

Also make a single column primary key a requirement for live tables. We
can't rely on all versioned tables having an "id" column as their
primary key, which is why we get the primary key column automatically.
Extending this to multi-column support shouldn't be hard, but is not
needed right now.