History tables: don't sync tables if history is enabled twice
If enable_history_tracking_for_table() is called with the history system
already being set up for a particular table, no sync will be performed
anymore. A sync is unnecessary in this situation, because history table
triggers should already take care of this. Therefore, a sync is now only
performed if the history triggers weren't set up before.